Move to a src layout
