--- server/__init__.py
@@ -1,5 +1,0 @@
-"""ZAPP Atlas server package.
-
-We keep this lightweight; its primary purpose is to make imports work from the
-repository root (e.g. `from server.api.main import app`).
-"""
