+++ server/pyproject.toml
@@ -1,7 +1,7 @@
 [project]
-name = "zapp-atlas-server"
+name = "zapp-atlas"
 version = "0.1.0"
-description = "Server for the ZAPP atlas"
+description = "Server and utilities for the ZAPP atlas"
 readme = "README.md"
 requires-python = ">=3.12"
 dependencies = [
@@ -13,11 +13,22 @@ dependencies = [
     "zebrafish-toxicology-atlas-schema @ git+https://github.com/zappfish/zebrafish-toxicology-atlas-schema.git@main",
 ]
 
+# FIXME: Remove this after moving schema inline
+[tool.hatch.metadata]
+allow-direct-references = true
+
 [dependency-groups]
 dev = [
     "httpx>=0.27.0",
     "pytest>=8.0",
 ]
 
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/zapp_atlas"]
+
 [tool.pytest.ini_options]
-pythonpath = [".."]
+pythonpath = ["src"]
