+++ server/src/zapp_atlas/api/routers/images.py
@@ -20,8 +20,8 @@ from fastapi import (
 from fastapi.responses import RedirectResponse, Response
 from sqlalchemy.orm import Session
 
-from server.api.deps import get_session
-from server.api.services.images import (
+from zapp_atlas.api.deps import get_session
+from zapp_atlas.api.services.images import (
     ImageTooLargeError,
     UnsupportedImageTypeError,
     create_image_for_observation,
@@ -30,7 +30,7 @@ from server.api.services.images import (
     image_url,
     load_image_bytes,
 )
-from server.storage import Storage, get_storage
+from zapp_atlas.db.image_storage import Storage, get_storage
 
 from zebrafish_toxicology_atlas_schema.datamodel.pydanticmodel_v2 import ImageRead
