+++ server/src/zapp_atlas/api/routers/observations.py
@@ -12,14 +12,14 @@ from typing import Annotated
 from fastapi import APIRouter, Depends, HTTPException, status
 from sqlalchemy.orm import Session
 
-from server.api.deps import get_session
-from server.api.services.observations import (
+from zapp_atlas.api.deps import get_session
+from zapp_atlas.api.services.observations import (
     create_observation_for_exposure,
     delete_observation,
     get_observation_by_id,
     patch_observation,
 )
-from server.storage import Storage, get_storage
+from zapp_atlas.db.image_storage import Storage, get_storage
 
 from zebrafish_toxicology_atlas_schema.datamodel.pydanticmodel_v2 import (
     PhenotypeObservationSetCreate,
