+++ server/src/zapp_atlas/api/routers/studies.py
@@ -11,15 +11,15 @@ from typing import Annotated
 from fastapi import APIRouter, Depends, HTTPException, status
 from sqlalchemy.orm import Session
 
-from server.api.deps import get_session
-from server.api.services.studies import (
+from zapp_atlas.api.deps import get_session
+from zapp_atlas.api.services.studies import (
     create_study,
     delete_study,
     get_study_by_id,
     list_studies,
     patch_study,
 )
-from server.storage import Storage, get_storage
+from zapp_atlas.db.image_storage import Storage, get_storage
 
 # LinkML-generated Pydantic CRUD models
 from zebrafish_toxicology_atlas_schema.datamodel.pydanticmodel_v2 import (
