+++ server/src/zapp_atlas/db/__init__.py
@@ -1,0 +1,2 @@
+from .db import *
+from .init_db import *
