+++ server/src/zapp_atlas/db/image_storage.py
