+++ server/src/zapp_atlas/main.py
@@ -11,13 +11,13 @@ from contextlib import asynccontextmanager
 
 from fastapi import APIRouter, FastAPI
 
-from server.api.routers.experiments import router as experiments_router
-from server.api.routers.exposures import router as exposures_router
-from server.api.routers.images import router as images_router
-from server.api.routers.observations import router as observations_router
-from server.api.routers.studies import router as studies_router
-from server.db import get_engine, get_session_factory, init_db
-from server.seed import seed
+from zapp_atlas.api.routers.experiments import router as experiments_router
+from zapp_atlas.api.routers.exposures import router as exposures_router
+from zapp_atlas.api.routers.images import router as images_router
+from zapp_atlas.api.routers.observations import router as observations_router
+from zapp_atlas.api.routers.studies import router as studies_router
+from zapp_atlas.db import get_engine, get_session_factory, init_db
+from zapp_atlas.seed import seed
 
 
 @asynccontextmanager
