+++ server/src/zapp_atlas/seed.py
@@ -21,7 +21,7 @@ from __future__ import annotations
 
 from sqlalchemy.orm import Session
 
-from server.db import get_engine, get_session_factory, init_db
+from zapp_atlas.db import get_engine, get_session_factory, init_db
 from zebrafish_toxicology_atlas_schema.datamodel.sqla import (  # type: ignore
     ExposureEvent,
     ExposureRoute,
