+++ server/tests/test_api_experiments.py
@@ -7,8 +7,8 @@ from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy.pool import StaticPool
 
-from server.api.deps import get_session
-from server.api.main import create_app
+from zapp_atlas.api.deps import get_session
+from zapp_atlas.main import create_app
 
 
 def _make_test_app():
